refactor: 승인, 취소가능 여부 service로 이관
