+++ src/test/java/nextstep/courses/domain/EnrollmentTest.java
@@ -16,19 +16,10 @@ public class EnrollmentTest {
     public static final Enrollment ENROLLMENT2 = Enrollment.enrollToGeneralCourse(1L, NsUserTest.SANJIGI, new Payment());
     public static final Enrollments ENROLLMENTS = new Enrollments(new ArrayList<>(List.of(ENROLLMENT1, ENROLLMENT2)));
 
-    @Test
-    void 강사만_승인_거절_가능() {
-        Enrollment enrollment = Enrollment.enrollToSelectedCourse(1L, NsUserTest.JAVAJIGI, new Payment());
-        enrollment.approve(NsUserTest.INSTRUCTOR);
-        Assertions.assertThat(enrollment.getStatus()).isEqualTo(EnrollmentStatus.APPROVED);
-    }
 
     @Test
-    void 강사_아니면_승인_거절_예외() {
+    void 신청상태만_승인가능() {
         Assertions.assertThatIllegalArgumentException()
-                .isThrownBy(() -> ENROLLMENT1.approve(NsUserTest.SANJIGI));
-        Assertions.assertThatIllegalArgumentException()
-                .isThrownBy(() -> ENROLLMENT1.cancel(NsUserTest.SANJIGI));
-
+                .isThrownBy(ENROLLMENT1::approve);
     }
 }
